+++ crates/objc2/src/runtime/protocol_object.rs
@@ -7,7 +7,7 @@ use crate::encode::{Encoding, RefEncode};
 use crate::rc::{autoreleasepool_leaking, Id};
 use crate::runtime::__nsstring::nsstring_to_str;
 use crate::runtime::{AnyObject, NSObjectProtocol};
-use crate::{Message, ProtocolType};
+use crate::Message;
 
 /// An internal helper trait for [`ProtocolObject`].
 ///
@@ -57,25 +57,35 @@ pub unsafe trait ImplementedBy<T: ?Sized + Message> {
 /// ```
 #[doc(alias = "id")]
 #[repr(C)]
-pub struct ProtocolObject<P: ?Sized + ProtocolType> {
+pub struct ProtocolObject<P: ?Sized> {
     inner: AnyObject,
     p: PhantomData<P>,
 }
 
+// SAFETY: `Send` if the underlying trait promises `Send`.
+//
+// E.g. `ProtocolObject<dyn NSObjectProtocol + Send>` is naturally `Send`.
+unsafe impl<P: ?Sized + Send> Send for ProtocolObject<P> {}
+
+// SAFETY: `Sync` if the underlying trait promises `Sync`.
+//
+// E.g. `ProtocolObject<dyn NSObjectProtocol + Sync>` is naturally `Sync`.
+unsafe impl<P: ?Sized + Sync> Sync for ProtocolObject<P> {}
+
 // SAFETY: The type is `#[repr(C)]` and `AnyObject` internally
-unsafe impl<P: ?Sized + ProtocolType> RefEncode for ProtocolObject<P> {
+unsafe impl<P: ?Sized> RefEncode for ProtocolObject<P> {
     const ENCODING_REF: Encoding = Encoding::Object;
 }
 
 // SAFETY: The type is `AnyObject` internally, and is mean to be messaged
 // as-if it's an object.
-unsafe impl<P: ?Sized + ProtocolType> Message for ProtocolObject<P> {}
+unsafe impl<P: ?Sized> Message for ProtocolObject<P> {}
 
-impl<P: ?Sized + ProtocolType> ProtocolObject<P> {
+impl<P: ?Sized> ProtocolObject<P> {
     /// Get an immutable type-erased reference from a type implementing a
     /// protocol.
     #[inline]
-    pub fn from_ref<T: Message>(obj: &T) -> &Self
+    pub fn from_ref<T: ?Sized + Message>(obj: &T) -> &Self
     where
         P: ImplementedBy<T>,
     {
@@ -89,7 +99,7 @@ impl<P: ?Sized + ProtocolType> ProtocolObject<P> {
     /// Get a mutable type-erased reference from a type implementing a
     /// protocol.
     #[inline]
-    pub fn from_mut<T: Message>(obj: &mut T) -> &mut Self
+    pub fn from_mut<T: ?Sized + Message>(obj: &mut T) -> &mut Self
     where
         P: ImplementedBy<T>,
     {
@@ -118,7 +128,7 @@ impl<P: ?Sized + ProtocolType> ProtocolObject<P> {
     }
 }
 
-impl<P: ?Sized + ProtocolType + NSObjectProtocol> PartialEq for ProtocolObject<P> {
+impl<P: ?Sized + NSObjectProtocol> PartialEq for ProtocolObject<P> {
     #[inline]
     #[doc(alias = "isEqual:")]
     fn eq(&self, other: &Self) -> bool {
@@ -126,16 +136,16 @@ impl<P: ?Sized + ProtocolType + NSObjectProtocol> PartialEq for ProtocolObject<P
     }
 }
 
-impl<P: ?Sized + ProtocolType + NSObjectProtocol> Eq for ProtocolObject<P> {}
+impl<P: ?Sized + NSObjectProtocol> Eq for ProtocolObject<P> {}
 
-impl<P: ?Sized + ProtocolType + NSObjectProtocol> hash::Hash for ProtocolObject<P> {
+impl<P: ?Sized + NSObjectProtocol> hash::Hash for ProtocolObject<P> {
     #[inline]
     fn hash<H: hash::Hasher>(&self, state: &mut H) {
         self.__hash().hash(state);
     }
 }
 
-impl<P: ?Sized + ProtocolType + NSObjectProtocol> fmt::Debug for ProtocolObject<P> {
+impl<P: ?Sized + NSObjectProtocol> fmt::Debug for ProtocolObject<P> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         // Attempt to format description string
         if let Some(description) = self.__description() {
@@ -159,10 +169,9 @@ impl<P: ?Sized + ProtocolType + NSObjectProtocol> fmt::Debug for ProtocolObject<
     }
 }
 
-impl<P, T> AsRef<ProtocolObject<T>> for ProtocolObject<P>
+impl<P: ?Sized, T> AsRef<ProtocolObject<T>> for ProtocolObject<P>
 where
-    P: ?Sized + ProtocolType,
-    T: ?Sized + ProtocolType + ImplementedBy<ProtocolObject<P>>,
+    T: ?Sized + ImplementedBy<ProtocolObject<P>>,
 {
     #[inline]
     fn as_ref(&self) -> &ProtocolObject<T> {
@@ -170,10 +179,9 @@ where
     }
 }
 
-impl<P, T> AsMut<ProtocolObject<T>> for ProtocolObject<P>
+impl<P: ?Sized, T> AsMut<ProtocolObject<T>> for ProtocolObject<P>
 where
-    P: ?Sized + ProtocolType,
-    T: ?Sized + ProtocolType + ImplementedBy<ProtocolObject<P>>,
+    T: ?Sized + ImplementedBy<ProtocolObject<P>>,
 {
     #[inline]
     fn as_mut(&mut self) -> &mut ProtocolObject<T> {
@@ -181,7 +189,7 @@ where
     }
 }
 
-// TODO: Maybe iplement Borrow + BorrowMut?
+// TODO: Maybe implement Borrow + BorrowMut?
 
 #[cfg(test)]
 #[allow(clippy::missing_safety_doc)]
@@ -194,7 +202,9 @@ mod tests {
     use super::*;
     use crate::mutability::Mutable;
     use crate::runtime::{NSObject, NSObjectProtocol};
-    use crate::{declare_class, extern_methods, extern_protocol, ClassType, DeclaredClass};
+    use crate::{
+        declare_class, extern_methods, extern_protocol, ClassType, DeclaredClass, ProtocolType,
+    };
 
     extern_protocol!(
         unsafe trait Foo {
@@ -264,6 +274,9 @@ mod tests {
     unsafe impl FooBar for DummyClass {}
     // unsafe impl FooFooBar for DummyClass {}
 
+    unsafe impl Send for DummyClass {}
+    unsafe impl Sync for DummyClass {}
+
     extern_methods!(
         unsafe impl DummyClass {
             #[method_id(new)]
@@ -275,6 +288,12 @@ mod tests {
     fn impl_traits() {
         assert_impl_all!(NSObject: NSObjectProtocol);
         assert_impl_all!(ProtocolObject<dyn NSObjectProtocol>: NSObjectProtocol);
+        assert_not_impl_any!(ProtocolObject<dyn NSObjectProtocol>: Send, Sync);
+        assert_impl_all!(ProtocolObject<dyn NSObjectProtocol + Send>: NSObjectProtocol, Send);
+        assert_not_impl_any!(ProtocolObject<dyn NSObjectProtocol + Send>: Sync);
+        assert_impl_all!(ProtocolObject<dyn NSObjectProtocol + Sync>: NSObjectProtocol, Sync);
+        assert_not_impl_any!(ProtocolObject<dyn NSObjectProtocol + Sync>: Send);
+        assert_impl_all!(ProtocolObject<dyn NSObjectProtocol + Send + Sync>: NSObjectProtocol, Send, Sync);
         assert_not_impl_any!(ProtocolObject<dyn Foo>: NSObjectProtocol);
         assert_impl_all!(ProtocolObject<dyn Bar>: NSObjectProtocol);
         assert_impl_all!(ProtocolObject<dyn FooBar>: NSObjectProtocol);
@@ -332,6 +351,10 @@ mod tests {
         let _nsobject: &ProtocolObject<dyn NSObjectProtocol> = ProtocolObject::from_ref(bar);
         let nsobject: &ProtocolObject<dyn NSObjectProtocol> = ProtocolObject::from_ref(&*obj);
         let _nsobject: &ProtocolObject<dyn NSObjectProtocol> = ProtocolObject::from_ref(nsobject);
+        let _: &ProtocolObject<dyn NSObjectProtocol + Send> = ProtocolObject::from_ref(&*obj);
+        let _: &ProtocolObject<dyn NSObjectProtocol + Sync> = ProtocolObject::from_ref(&*obj);
+        let _: &ProtocolObject<dyn NSObjectProtocol + Send + Sync> =
+            ProtocolObject::from_ref(&*obj);
 
         let _foobar: &mut ProtocolObject<dyn FooBar> = ProtocolObject::from_mut(&mut *obj);
         let _foobar: Id<ProtocolObject<dyn FooBar>> = ProtocolObject::from_id(obj);